Added / moved all classes related to view quality to ViewQuality folder in Humans_Om. Some logical renaming
